+++ src/core/internal/delegation.test.ts
@@ -6,7 +6,7 @@ import { describe, expect, test } from 'vitest'
 import { getAccount } from '../../../test/src/account.js'
 import { client, delegation } from '../../../test/src/porto.js'
 import * as Call from './call.js'
-import * as DelegatedAccount from './delegatedAccount.js'
+import * as Delegation from './delegation.js'
 import * as Key from './key.js'
 
 describe('execute', () => {
@@ -18,7 +18,7 @@ describe('execute', () => {
         role: 'admin',
       })
 
-      await DelegatedAccount.execute(client, {
+      await Delegation.execute(client, {
         account,
         calls: [
           Call.authorize({
@@ -29,7 +29,7 @@ describe('execute', () => {
       })
 
       expect(
-        await DelegatedAccount.keyAt(client, {
+        await Delegation.keyAt(client, {
           account,
           index: 0,
         }),
@@ -44,7 +44,7 @@ describe('execute', () => {
         delegation,
       })
 
-      await DelegatedAccount.execute(client, {
+      await Delegation.execute(client, {
         account,
         calls: [],
         initialize: true,
@@ -54,7 +54,7 @@ describe('execute', () => {
         role: 'admin',
       })
 
-      await DelegatedAccount.execute(client, {
+      await Delegation.execute(client, {
         account,
         calls: [
           Call.authorize({
@@ -64,7 +64,7 @@ describe('execute', () => {
       })
 
       expect(
-        await DelegatedAccount.keyAt(client, {
+        await Delegation.keyAt(client, {
           account,
           index: 0,
         }),
@@ -83,7 +83,7 @@ describe('execute', () => {
         role: 'admin',
       })
 
-      await DelegatedAccount.execute(client, {
+      await Delegation.execute(client, {
         account,
         calls: [
           Call.authorize({
@@ -95,7 +95,7 @@ describe('execute', () => {
       })
 
       expect(
-        await DelegatedAccount.keyAt(client, {
+        await Delegation.keyAt(client, {
           account,
           index: 0,
         }),
@@ -110,7 +110,7 @@ describe('execute', () => {
         delegation,
       })
 
-      await DelegatedAccount.execute(client, {
+      await Delegation.execute(client, {
         account,
         calls: [],
         initialize: true,
@@ -120,7 +120,7 @@ describe('execute', () => {
         role: 'admin',
       })
 
-      await DelegatedAccount.execute(client, {
+      await Delegation.execute(client, {
         account,
         calls: [
           Call.authorize({
@@ -131,7 +131,7 @@ describe('execute', () => {
       })
 
       expect(
-        await DelegatedAccount.keyAt(client, {
+        await Delegation.keyAt(client, {
           account,
           index: 0,
         }),
@@ -148,7 +148,7 @@ describe('execute', () => {
         role: 'admin',
       })
 
-      await DelegatedAccount.execute(client, {
+      await Delegation.execute(client, {
         account,
         calls: [
           Call.authorize({
@@ -163,7 +163,7 @@ describe('execute', () => {
         role: 'admin',
       })
 
-      await DelegatedAccount.execute(client, {
+      await Delegation.execute(client, {
         account,
         calls: [
           Call.authorize({
@@ -174,7 +174,7 @@ describe('execute', () => {
       })
 
       expect(
-        await DelegatedAccount.keyAt(client, {
+        await Delegation.keyAt(client, {
           account,
           index: 1,
         }),
@@ -191,7 +191,7 @@ describe('execute', () => {
         role: 'admin',
       })
 
-      await DelegatedAccount.execute(client, {
+      await Delegation.execute(client, {
         account,
         calls: [
           Call.authorize({
@@ -206,7 +206,7 @@ describe('execute', () => {
         role: 'admin',
       })
 
-      await DelegatedAccount.execute(client, {
+      await Delegation.execute(client, {
         account,
         calls: [
           Call.authorize({
@@ -217,7 +217,7 @@ describe('execute', () => {
       })
 
       expect(
-        await DelegatedAccount.keyAt(client, {
+        await Delegation.keyAt(client, {
           account,
           index: 1,
         }),
@@ -234,7 +234,7 @@ describe('execute', () => {
         delegation,
       })
 
-      await DelegatedAccount.execute(client, {
+      await Delegation.execute(client, {
         account,
         calls: [],
         initialize: true,
@@ -252,7 +252,7 @@ describe('execute', () => {
       expect(balances_before[1]).toEqual(Value.fromEther('0'))
       expect(balances_before[2]).toEqual(Value.fromEther('0'))
 
-      await DelegatedAccount.execute(client, {
+      await Delegation.execute(client, {
         account,
         calls: [
           { to: alice.address, value: Value.fromEther('1') },
@@ -278,7 +278,7 @@ describe('execute', () => {
         delegation,
       })
 
-      await DelegatedAccount.execute(client, {
+      await Delegation.execute(client, {
         account,
         calls: [],
         initialize: true,
@@ -296,7 +296,7 @@ describe('execute', () => {
       expect(balances_before[1]).toEqual(Value.fromEther('0'))
       expect(balances_before[2]).toEqual(Value.fromEther('0'))
 
-      await DelegatedAccount.execute(client, {
+      await Delegation.execute(client, {
         account,
         calls: [
           { to: alice.address, value: Value.fromEther('1') },
@@ -320,46 +320,6 @@ describe('execute', () => {
   })
 })
 
-describe('from', () => {
-  test('default', () => {
-    const account = DelegatedAccount.from({
-      address: '0x0000000000000000000000000000000000000001',
-      delegation: '0x0000000000000000000000000000000000000002',
-      keys: [
-        Key.fromP256({
-          expiry: 42069,
-          privateKey:
-            '0x59ff6b8de3b3b39e94b6f9fc0590cf4e3eaa9b6736e6a49c9a6b324c4f58cb9f',
-          role: 'admin',
-        }),
-      ],
-      label: 'test',
-    })
-
-    expect(account).toMatchInlineSnapshot(`
-      {
-        "address": "0x0000000000000000000000000000000000000001",
-        "delegation": "0x0000000000000000000000000000000000000002",
-        "keys": [
-          {
-            "expiry": 42069,
-            "publicKey": {
-              "prefix": 4,
-              "x": 106772332543853129808885768556237579026047721040871197501406793325324706664735n,
-              "y": 110339676533900797749866890631856969312064778492080444630962038345769628880904n,
-            },
-            "role": "admin",
-            "sign": [Function],
-            "type": "p256",
-          },
-        ],
-        "label": "test",
-        "type": "delegated",
-      }
-    `)
-  })
-})
-
 describe('prepareExecute', () => {
   describe('authorize', () => {
     test('delegated: false, key: owner, keysToAuthorize: [P256], executor: JSON-RPC', async () => {
@@ -371,7 +331,7 @@ describe('prepareExecute', () => {
         role: 'admin',
       })
 
-      const { request, signPayloads } = await DelegatedAccount.prepareExecute(
+      const { request, signPayloads } = await Delegation.prepareExecute(
         client,
         {
           account,
@@ -388,13 +348,13 @@ describe('prepareExecute', () => {
         signPayloads.map((payload) => key.sign({ payload })),
       )
 
-      await DelegatedAccount.execute(client, {
+      await Delegation.execute(client, {
         ...request,
         signatures,
       })
 
       expect(
-        await DelegatedAccount.keyAt(client, {
+        await Delegation.keyAt(client, {
           account,
           index: 0,
         }),
@@ -409,7 +369,7 @@ describe('prepareExecute', () => {
         delegation,
       })
 
-      await DelegatedAccount.execute(client, {
+      await Delegation.execute(client, {
         account,
         calls: [],
         initialize: true,
@@ -421,7 +381,7 @@ describe('prepareExecute', () => {
         role: 'admin',
       })
 
-      const { request, signPayloads } = await DelegatedAccount.prepareExecute(
+      const { request, signPayloads } = await Delegation.prepareExecute(
         client,
         {
           account,
@@ -437,13 +397,13 @@ describe('prepareExecute', () => {
         signPayloads.map((payload) => key.sign({ payload })),
       )
 
-      await DelegatedAccount.execute(client, {
+      await Delegation.execute(client, {
         ...request,
         signatures,
       })
 
       expect(
-        await DelegatedAccount.keyAt(client, {
+        await Delegation.keyAt(client, {
           account,
           index: 0,
         }),
@@ -464,7 +424,7 @@ describe('prepareExecute', () => {
         role: 'admin',
       })
 
-      const { request, signPayloads } = await DelegatedAccount.prepareExecute(
+      const { request, signPayloads } = await Delegation.prepareExecute(
         client,
         {
           account,
@@ -482,13 +442,13 @@ describe('prepareExecute', () => {
         signPayloads.map((payload) => key.sign({ payload })),
       )
 
-      await DelegatedAccount.execute(client, {
+      await Delegation.execute(client, {
         ...request,
         signatures,
       })
 
       expect(
-        await DelegatedAccount.keyAt(client, {
+        await Delegation.keyAt(client, {
           account,
           index: 0,
         }),
@@ -508,7 +468,7 @@ describe('getExecuteSignPayload', () => {
       role: 'admin',
     })
 
-    const payload = await DelegatedAccount.getExecuteSignPayload(client, {
+    const payload = await Delegation.getExecuteSignPayload(client, {
       account,
       calls: [
         Call.authorize({
@@ -526,7 +486,7 @@ describe('getExecuteSignPayload', () => {
       delegation,
     })
 
-    await DelegatedAccount.execute(client, {
+    await Delegation.execute(client, {
       account,
       calls: [],
       initialize: true,
@@ -536,7 +496,7 @@ describe('getExecuteSignPayload', () => {
       role: 'admin',
     })
 
-    const payload = await DelegatedAccount.getExecuteSignPayload(client, {
+    const payload = await Delegation.getExecuteSignPayload(client, {
       account,
       calls: [
         Call.authorize({
