+++ src/core/internal/provider.test.ts
@@ -1,8 +1,20 @@
-import { expect, test } from 'vitest'
+import { describe, expect, test } from 'vitest'
 
 import { delegation, porto } from '../../../test/src/porto.js'
 
-test('default', async () => {
+describe('eth_accounts', () => {
+  test('behavior: disconnected', async () => {
+    await expect(
+      porto.provider.request({
+        method: 'eth_accounts',
+      }),
+    ).rejects.toThrowErrorMatchingInlineSnapshot(
+      '[Provider.DisconnectedError: The provider is disconnected from all chains.]',
+    )
+  })
+})
+
+test('smoke', async () => {
   const code = await porto.provider.request({
     method: 'eth_getCode',
     params: [delegation, 'latest'],
